Adjust spacing and dividers for text categories
RISDEV-4729
