+++ frontend/src/components/texts/DocumentUnitLongTexts.vue
@@ -33,24 +33,6 @@ const hasDecisionReasons = ref<boolean>(
     : false,
 )
 
-const hasOtherLongtext = ref<boolean>(
-  store.documentUnit?.longTexts?.otherLongText
-    ? store.documentUnit?.longTexts?.otherLongText?.length > 0
-    : false,
-)
-
-const hasDissentingOpinion = ref<boolean>(
-  store.documentUnit?.longTexts?.dissentingOpinion
-    ? store.documentUnit?.longTexts?.dissentingOpinion?.length > 0
-    : false,
-)
-
-const hasOutline = ref<boolean>(
-  store.documentUnit?.longTexts?.outline
-    ? store.documentUnit?.longTexts?.outline?.length > 0
-    : false,
-)
-
 const tenor = computed({
   get: () => store.documentUnit?.longTexts.tenor,
   set: (newValue) => {
@@ -87,33 +69,6 @@ const decisionReasons = computed({
   },
 })
 
-const otherLongText = computed({
-  get: () => store.documentUnit?.longTexts.otherLongText,
-  set: (newValue) => {
-    store.documentUnit!.longTexts.otherLongText = hasContent(newValue)
-      ? newValue
-      : undefined
-  },
-})
-
-const dissentingOpinion = computed({
-  get: () => store.documentUnit?.longTexts.dissentingOpinion,
-  set: (newValue) => {
-    store.documentUnit!.longTexts.dissentingOpinion = hasContent(newValue)
-      ? newValue
-      : undefined
-  },
-})
-
-const outline = computed({
-  get: () => store.documentUnit?.longTexts.outline,
-  set: (newValue) => {
-    store.documentUnit!.longTexts.outline = hasContent(newValue)
-      ? newValue
-      : undefined
-  },
-})
-
 function hasContent(text?: string) {
   const divElem = document.createElement("div")
   if (text == undefined) text = ""
@@ -126,78 +81,47 @@ function hasContent(text?: string) {
 </script>
 
 <template>
-  <h2 class="ds-heading-03-bold mb-16 mt-24">Langtexte</h2>
-  <div class="flex flex-col gap-24">
-    <CategoryWrapper label="Tenor" :should-show-button="!hasTenor">
-      <DocumentUnitTextField
-        id="tenor"
-        v-model="tenor"
-        :editable="isInternalUser"
-        label="Tenor"
-      />
-    </CategoryWrapper>
+  <div aria-label="Langtexte" class="border-b-1 border-blue-300 pb-24">
+    <h2 class="ds-label-01-bold mb-16 mt-24">Langtexte</h2>
+    <div class="flex flex-col gap-24">
+      <CategoryWrapper label="Tenor" :should-show-button="!hasTenor">
+        <DocumentUnitTextField
+          id="tenor"
+          v-model="tenor"
+          :editable="isInternalUser"
+          label="Tenor"
+        />
+      </CategoryWrapper>
 
-    <CategoryWrapper label="Gründe" :should-show-button="!hasReasons">
-      <DocumentUnitTextField
-        id="reasons"
-        v-model="reasons"
-        :editable="isInternalUser"
-        label="Gründe"
-      />
-    </CategoryWrapper>
+      <CategoryWrapper label="Gründe" :should-show-button="!hasReasons">
+        <DocumentUnitTextField
+          id="reasons"
+          v-model="reasons"
+          :editable="isInternalUser"
+          label="Gründe"
+        />
+      </CategoryWrapper>
 
-    <CategoryWrapper label="Tatbestand" :should-show-button="!hasCaseFacts">
-      <DocumentUnitTextField
-        id="caseFacts"
-        v-model="caseFacts"
-        :editable="isInternalUser"
-        label="Tatbestand"
-      />
-    </CategoryWrapper>
+      <CategoryWrapper label="Tatbestand" :should-show-button="!hasCaseFacts">
+        <DocumentUnitTextField
+          id="caseFacts"
+          v-model="caseFacts"
+          :editable="isInternalUser"
+          label="Tatbestand"
+        />
+      </CategoryWrapper>
 
-    <CategoryWrapper
-      label="Entscheidungsgründe"
-      :should-show-button="!hasDecisionReasons"
-    >
-      <DocumentUnitTextField
-        id="decisionReasons"
-        v-model="decisionReasons"
-        :editable="isInternalUser"
+      <CategoryWrapper
         label="Entscheidungsgründe"
-      />
-    </CategoryWrapper>
-
-    <CategoryWrapper
-      label="Sonstiger Langtext"
-      :should-show-button="!hasOtherLongtext"
-    >
-      <DocumentUnitTextField
-        id="otherLongText"
-        v-model="otherLongText"
-        :editable="isInternalUser"
-        label="Sonstiger Langtext"
-      />
-    </CategoryWrapper>
-
-    <CategoryWrapper
-      label="Abweichende Meinung"
-      :should-show-button="!hasDissentingOpinion"
-    >
-      <DocumentUnitTextField
-        id="dissentingOpinion"
-        v-model="dissentingOpinion"
-        :editable="isInternalUser"
-        label="Abweichende Meinung"
-      />
-    </CategoryWrapper>
-
-    <CategoryWrapper label="Gliederung" :should-show-button="!hasOutline">
-      <DocumentUnitTextField
-        id="outline"
-        v-model="outline"
-        :editable="isInternalUser"
-        label="Gliederung"
-      />
-    </CategoryWrapper>
+        :should-show-button="!hasDecisionReasons"
+      >
+        <DocumentUnitTextField
+          id="decisionReasons"
+          v-model="decisionReasons"
+          :editable="isInternalUser"
+          label="Entscheidungsgründe"
+        />
+      </CategoryWrapper>
+    </div>
   </div>
 </template>
